requirements.txt
pandas==1.2.4
mysqlclient==2.0.3